+++ src/utn.h
@@ -11,6 +11,6 @@ int getNumber(int *pResult, char *message, char *errorMessage, int attemps);
 int add(int* pResult, int firstNumber, int secondNumber);
 int subtract(int* pResult, int firstNumber, int secondNumber);
 int multiply(int* pResult, int firstNumber, int secondNumber);
-int division(float* pResult, int divisor, int dividend);
+int division(int* pResult, int divisor, int dividend);
 int factorial(int *pResult, int operator);
 #endif /* SRC_LIB_H_ */
